Conditonal renderning for navigation based on authentication
