+++ components/NavBar.js
@@ -1,29 +1,46 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
 import React from 'react';
-import Link from 'next/link';
 import {
-  Navbar, Container, Nav, Button,
+  Navbar, Container, Nav,
 } from 'react-bootstrap';
-import { signOut } from '../utils/auth';
+import { BiBookReader } from 'react-icons/bi';
+import { MdOutlineCoffeeMaker } from 'react-icons/md';
+import { GrFavorite, GrUserSettings } from 'react-icons/gr';
+import { useAuth } from '../utils/context/authContext';
 
 export default function NavBar() {
+  const { user } = useAuth();
   return (
-    <Navbar collapseOnSelect expand="lg" bg="dark" variant="dark">
-      <Container>
-        <Link passHref href="/">
-          <Navbar.Brand>CHANGE ME</Navbar.Brand>
-        </Link>
-        <Navbar.Toggle aria-controls="responsive-navbar-nav" />
-        <Navbar.Collapse id="responsive-navbar-nav">
-          <Nav className="me-auto">
-            {/* CLOSE NAVBAR ON LINK SELECTION: https://stackoverflow.com/questions/72813635/collapse-on-select-react-bootstrap-navbar-with-nextjs-not-working */}
-            <Link passHref href="/">
-              <Nav.Link>Home</Nav.Link>
-            </Link>
-            <Button variant="danger" onClick={signOut}>Sign Out</Button>
-          </Nav>
-        </Navbar.Collapse>
-      </Container>
+    <Navbar fixed="bottom" collapseOnSelect expand="true" bg="transparent" variant="dark">
+      {
+        user
+          ? (
+            <Container>
+              <Nav.Link href="/">
+                <BiBookReader />
+              </Nav.Link>
+              <Nav.Link href="/">
+                <MdOutlineCoffeeMaker />
+              </Nav.Link>
+              <Nav.Link href="/">
+                <GrFavorite />
+              </Nav.Link>
+              <Nav.Link href="/settings">
+                <GrUserSettings />
+              </Nav.Link>
+            </Container>
+          )
+          : (
+            <Container>
+              <Nav.Link href="/">
+                <MdOutlineCoffeeMaker />
+              </Nav.Link>
+              <Nav.Link href="/settings">
+                <GrUserSettings />
+              </Nav.Link>
+            </Container>
+          )
+      }
     </Navbar>
   );
 }
