+++ components/Signin.js
@@ -1,25 +1,25 @@
-import React from 'react';
-import { Button } from 'react-bootstrap';
-import { signIn } from '../utils/auth';
+// import React from 'react';
+// import { Button } from 'react-bootstrap';
+// import { signIn } from '../utils/auth';
 
-function Signin() {
-  return (
-    <div
-      className="text-center d-flex flex-column justify-content-center align-content-center"
-      style={{
-        height: '90vh',
-        padding: '30px',
-        maxWidth: '400px',
-        margin: '0 auto',
-      }}
-    >
-      <h1>Hi there!</h1>
-      <p>Click the button below to login!</p>
-      <Button type="button" size="lg" className="copy-btn" onClick={signIn}>
-        Sign In
-      </Button>
-    </div>
-  );
-}
+// function Signin() {
+//   return (
+//     <div
+//       className="text-center d-flex flex-column justify-content-center align-content-center"
+//       style={{
+//         height: '90vh',
+//         padding: '30px',
+//         maxWidth: '400px',
+//         margin: '0 auto',
+//       }}
+//     >
+//       <h1>Hi there!</h1>
+//       <p>Click the button below to login!</p>
+//       <Button type="button" size="lg" className="copy-btn" onClick={signIn}>
+//         Sign In
+//       </Button>
+//     </div>
+//   );
+// }
 
-export default Signin;
+// export default Signin;
