+++ package.json
@@ -25,7 +25,8 @@
     "prop-types": "^15.7.2",
     "react": "17.0.2",
     "react-bootstrap": "^2.4.0",
-    "react-dom": "17.0.2"
+    "react-dom": "17.0.2",
+    "react-icons": "^4.4.0"
   },
   "devDependencies": {
     "eslint": "^7.32.0",
