--- pages/delete-me.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function DeleteMe() {
-  return <div>DeleteMe</div>;
-}
