+++ pages/recipes/favRecipes.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function favRecipes() {
+  return (
+    <div>favRecipes</div>
+  );
+}
