+++ pages/recipes/userRecipes.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function userRecipes() {
+  return (
+    <div>userRecipes</div>
+  );
+}
