+++ utils/ViewDirector.js
@@ -5,7 +5,7 @@ import Signin from '../components/Signin';
 import NavBar from '../components/NavBar';
 
 const ViewDirectorBasedOnUserAuthStatus = ({ component: Component, pageProps }) => {
-  const { user, userLoading } = useAuth();
+  const { userLoading } = useAuth();
 
   // if user state is null, then show loader
   if (userLoading) {
@@ -13,18 +13,14 @@ const ViewDirectorBasedOnUserAuthStatus = ({ component: Component, pageProps }) 
   }
 
   // what the user should see if they are logged in
-  if (user) {
-    return (
-      <>
-        <NavBar /> {/* NavBar only visible if user is logged in and is in every view */}
-        <div className="container">
-          <Component {...pageProps} />
-        </div>
-      </>
-    );
-  }
-
-  return <Signin />;
+  return (
+    <>
+      <NavBar />
+      <div className="container">
+        <Component {...pageProps} />
+      </div>
+    </>
+  );
 };
 
 export default ViewDirectorBasedOnUserAuthStatus;
